requirements.txt
aiogram==2.20
aiohttp==3.7.4.post0
aiosignal==1.2.0
aiosocksy==0.1.2
aiotg==1.0.0
altgraph==0.17.3
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astroid==3.0.1
asttokens==2.0.8
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.4
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.7.0
bottle==0.12.23
bottle-websocket==0.2.9
cachetools==4.2.2
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.1.7
config==0.5.1
contourpy==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cycler==0.11.0
Cython==0.29.28
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.10
django-allauth==0.49.0
django-bootstrap4==23.2
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.14.0
django-grappelli==2.15.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
dnspython==2.4.2
docker==7.0.0
docutils==0.19
drawdata==0.2.0
drf-yasg==1.20.0
Eel==0.15.1
email-validator==2.1.0.post1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.1.1
fastapi==0.104.1
fastjsonschema==2.16.2
feedparser==6.0.8
filelock==3.4.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.38.0
frozenlist==1.3.0
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.4
httptools==0.6.1
httpx==0.25.1
idna==2.10
importlib-metadata==5.0.0
inflection==0.5.1
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isort==5.12.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.10
jsonify==0.5
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.16.0
Kivy==2.1.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
lxml==4.9.2
macholib==1.16.2
MarkupSafe==2.1.3
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
motor==2.5.1
multidict==5.1.0
mypy==1.7.0
mypy-extensions==1.0.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
openpyxl==3.0.9
orjson==3.9.10
outcome==1.2.0
packaging==21.2
pandas==1.5.2
pandocfilters==1.5.0
parameterized==0.8.1
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.2
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2021.11.23
platformdirs==2.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.2
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.6.0
pycparser==2.21
pydantic==2.5.1
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.3
pygame==2.5.2
Pygments==2.13.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyJWT==2.1.0
pylint==3.0.2
pymongo==3.12.0
pyodbc==5.0.1
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyQt6==6.3.0
PyQt6-Qt6==6.3.0
PyQt6-sip==13.3.1
pyrsistent==0.18.1
PySocks==1.7.1
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==20.6
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsstail==0.5.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
selenium==4.9.0
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.23
sqlparse==0.4.2
stack-data==0.5.1
starlette==0.27.0
telepotpro==13.3
terminado==0.16.0
tinycss2==1.2.1
toml==0.10.1
tomli==2.0.1
tomlkit==0.12.3
tor==1.0.0
tornado==6.2
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==3.10.0.2
tzlocal==2.1
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
vsearch==1.1.0
watchdog==2.1.5
watchfiles==0.21.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==12.0
Werkzeug==3.0.1
whichcraft==0.6.1
whitenoise==3.0
widgetsnbextension==4.0.5
wsproto==1.2.0
yarl==1.6.3
youtube-dl==2021.12.17
youtube-search-python==1.5.3
zipp==3.9.0
zope.event==4.6
zope.interface==5.5.2
